[ feat ] OptionButton 컴포넌트로 분리하여 isClicked state 각각 달아주기
