+++ src/HyeEun/filter/components/common/OptionButton/style.ts
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+export const St = {
+  OptionButton: styled.button<{ isactive: boolean }>`
+    border: 1px solid lightgrey;
+    background-color: ${({ isactive }) => (isactive ? "#afafaf" : "transparent")};
+  `,
+};
